+++ example/bin/example.dart
@@ -1,0 +1,6 @@
+import 'package:dartson_test/serializer.dart';
+
+void main() {
+  print(serializer.encode(MyClass()..name = 'test'));
+  print(serializer.decode<MyClass>('{"name":"test"}').name);
+}