+++ lib/src/type_transformer.dart
@@ -1,5 +1,5 @@
-/// The interface for creating new transformers.
-/// For a basic example on how to use this interface take a look at [DateTimeParser].
+/// The interface for creating new transformers. For a basic example on how to
+/// use this interface take a look at [DateTimeParser].
 abstract class TypeTransformer<S, T> {
   /// Receives the [value] of type [T] and returns a serializable result which
   /// will be passed into the JSON representation.